+++ src/Celerity/TypeForwarders.cs
@@ -38,9 +38,11 @@ using Celerity.Primitives;
 [assembly: TypeForwardedTo(typeof(UInt32Hasher))]
 [assembly: TypeForwardedTo(typeof(UInt64Hasher))]
 #pragma warning restore CS0618
+[assembly: TypeForwardedTo(typeof(UInt32IdentityHasher))]
 [assembly: TypeForwardedTo(typeof(UInt32Murmur3Hasher))]
 [assembly: TypeForwardedTo(typeof(UInt32WangHasher))]
 [assembly: TypeForwardedTo(typeof(UInt32WangNaiveHasher))]
+[assembly: TypeForwardedTo(typeof(UInt64IdentityHasher))]
 [assembly: TypeForwardedTo(typeof(UInt64Murmur3Hasher))]
 [assembly: TypeForwardedTo(typeof(UInt64WangHasher))]
 [assembly: TypeForwardedTo(typeof(UInt64WangNaiveHasher))]